pyproject.toml
[project]
name = "summarizator"
version = "0.1.0"
description = "Summarization Techniques Demonstrator"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "attrs==25.4.0",
    "bert-extractive-summarizer==0.10.1",
    "blinker==1.9.0",
    "blis==1.3.0",
    "cachetools==6.2.1",
    "catalogue==2.0.10",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cloudpathlib==0.23.0",
    "confection==0.1.5",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "cymem==2.0.11",
    "en-core-web-sm",
    "filelock==3.20.0",
    "fonttools==4.60.1",
    "fsspec==2025.9.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "hf-xet==1.1.10",
    "hiredis==3.3.0",
    "huggingface-hub==0.35.3",
    "idna==3.11",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "kiwisolver==1.4.9",
    "langcodes==3.5.0",
    "language-data==1.3.0",
    "marisa-trie==1.3.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "murmurhash==1.0.13",
    "narwhals==2.7.0",
    "networkx==3.5",
    "nltk==3.9.2",
    "numpy==2.3.3",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==11.3.0",
    "preshed==3.0.10",
    "protobuf==6.32.1",
    "pyarrow==21.0.0",
    "pydantic==2.12.0",
    "pydantic-core==2.41.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis==6.4.0",
    "referencing==0.36.2",
    "regex==2025.9.18",
    "requests==2.32.5",
    "rich==14.2.0",
    "rpds-py==0.27.1",
    "ruff>=0.14.1",
    "sacremoses==0.1.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "sentencepiece==0.2.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.3.1",
    "smmap==5.0.2",
    "spacy==3.8.7",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.1",
    "streamlit==1.50.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "thinc==8.3.6",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "toml==0.10.2",
    "torch==2.8.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "transformers==4.57.0",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "wrapt==1.17.3",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
